+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Enhancements
     * `black` is default fill for text
     * true class extending. Now `rect instanceOf Kinetic.Shape` will return true
+    * while dragging you can redraw layer that is not under drag. hit graph will be updated in this case
 
 ## 5.1.9 2014-01-09